kanto_app/requirements.txt
paho-mqtt
ditto.client
kuksa_client.grpc